dev-requirements.txt
pytest-ckan
pytest-factoryboy
pytest==4.6.5
